Refactor agentchat to separate base interfaces from implementations
