+++ python/packages/autogen-agentchat/src/autogen_agentchat/base/_base_chat_agent.py
@@ -2,59 +2,13 @@ from abc import ABC, abstractmethod
 from typing import List, Sequence
 
 from autogen_core.base import CancellationToken
-from autogen_core.components import FunctionCall, Image
-from autogen_core.components.models import FunctionExecutionResult
 from autogen_core.components.tools import Tool
-from pydantic import BaseModel
+
+from ..messages import ChatMessage
+from ._base_task import TaskResult, TaskRunner
 
 
-class BaseMessage(BaseModel):
-    """A base message."""
-
-    source: str
-    """The name of the agent that sent this message."""
-
-
-class TextMessage(BaseMessage):
-    """A text message."""
-
-    content: str
-    """The content of the message."""
-
-
-class MultiModalMessage(BaseMessage):
-    """A multimodal message."""
-
-    content: List[str | Image]
-    """The content of the message."""
-
-
-class ToolCallMessage(BaseMessage):
-    """A message containing a list of function calls."""
-
-    content: List[FunctionCall]
-    """The list of function calls."""
-
-
-class ToolCallResultMessage(BaseMessage):
-    """A message containing the results of function calls."""
-
-    content: List[FunctionExecutionResult]
-    """The list of function execution results."""
-
-
-class StopMessage(BaseMessage):
-    """A message requesting stop of a conversation."""
-
-    content: str
-    """The content for the stop message."""
-
-
-ChatMessage = TextMessage | MultiModalMessage | StopMessage | ToolCallMessage | ToolCallResultMessage
-"""A message used by agents in a team."""
-
-
-class BaseChatAgent(ABC):
+class BaseChatAgent(TaskRunner, ABC):
     """Base class for a chat agent that can participant in a team."""
 
     def __init__(self, name: str, description: str) -> None:
@@ -80,6 +34,12 @@ class BaseChatAgent(ABC):
     async def on_messages(self, messages: Sequence[ChatMessage], cancellation_token: CancellationToken) -> ChatMessage:
         """Handle incoming messages and return a response message."""
         ...
+
+    async def run(
+        self, task: str, *, source: str = "user", cancellation_token: CancellationToken | None = None
+    ) -> TaskResult:
+        # TODO: Implement this method.
+        raise NotImplementedError
 
 
 class BaseToolUseChatAgent(BaseChatAgent):
