+++ python/packages/autogen-agentchat/src/autogen_agentchat/teams/__init__.py
@@ -1,9 +1,8 @@
 from ._group_chat._round_robin_group_chat import RoundRobinGroupChat
 from ._group_chat._selector_group_chat import SelectorGroupChat
-from ._termination import MaxMessageTermination, StopMessageTermination, TerminationCondition, TextMentionTermination
+from ._terminations import MaxMessageTermination, StopMessageTermination, TextMentionTermination
 
 __all__ = [
-    "TerminationCondition",
     "MaxMessageTermination",
     "TextMentionTermination",
     "StopMessageTermination",
